Move SubstateId into scrypto crate
