+++ radix-engine/src/state_manager/staging.rs
@@ -2,8 +2,9 @@ use sbor::rust::collections::*;
 use sbor::rust::vec::Vec;
 use scrypto::buffer::{scrypto_decode, scrypto_encode};
 
-use crate::{engine::SubstateId, ledger::*};
+use crate::{ledger::*};
 use indexmap::IndexMap;
+use scrypto::engine::types::SubstateId;
 
 /// Nodes form an acyclic graph towards the parent
 struct StagedSubstateStoreNode {
